refactor: inline query-params back into utils/index.ts
